app/requirements.txt
psycopg2-binary==2.9.6
kafka-python==2.0.2
pandas==1.0.5
numpy==1.19.5
sqlparse==0.4.4
neo4j==5.8.1
python-dotenv==1.0.0
SQLAlchemy==1.4.48
uvicorn==0.22.0
fastapi==0.95.2
google-cloud-secret-manager==2.16.1
py-bcrypt==0.4